feat(reviews): show comments on proposal review page
Reviewers couldn't comment on a proposal while reviewing it because
the review pane only rendered the proposal preview. Extract the
comments block into a shared ProposalComments component so both the
standalone proposal view and the review pane share the same UI.
